refactor(library): move sync_to_library + enqueue_write onto Library
Per review: free helpers `sync_to_library` and `enqueue_write` (which both
needed `&Library` or its inbox/search) move onto `impl Library` so they
read `self.inbox` / `self.search` directly instead of having them
threaded through arguments.

- `Library::sync_entity_in_op<E, OP>` replaces `sync_to_library<E, OP>`.
  Per-repo hooks call `self.library.sync_entity_in_op(op, entity, &mut events)`;
  the `Option<&Library>` handling moves to the (test-only) `Option<Library>`
  callers via a 4-line if-let.
- `Library::enqueue_write` is private. `Library::write_in_op`,
  `sync_workspace_folder_in_op`, `cleanup_workspace_in_op` use it directly.
