+++ core/src/workflow/repo.rs
@@ -56,6 +56,10 @@ impl WorkflowDefinitionRepo {
         entity: &WorkflowDefinition,
         mut new_events: es_entity::LastPersisted<'_, WorkflowDefinitionEvent>,
     ) -> Result<(), crate::library::LibraryError> {
-        crate::library::sync_to_library(self.library.as_ref(), op, entity, &mut new_events).await
+        if let Some(library) = &self.library {
+            library.sync_entity_in_op(op, entity, &mut new_events).await
+        } else {
+            Ok(())
+        }
     }
 }
